placed most oeprations in Ext trait
